Fix Dash immediately impacting ground
Dash was checking for impact on the client side which was originally at
eye height and now at ground level, resulting in immediate collision
with the block under the player; i.e. the original code in Dash works
fine, but the targeting algorithm needed to be adjusted.

Solution was to add half the searching entity's height to the y
position, but will need to keep an eye on this to make sure it doesn't
break other things.