Resolves issue #26 - fixed Mortal Draw failing to activate from a skill item
If player doesn't know Mortal Draw, the stack MUST be a provider, in which case the activation packet will trigger a re-scan for the ISkillProvider in order to ensure getPlayerSkill is able to return the skill.

In all cases, the first eligible item will be used, even if there may be a higher level one available. If the player knows the Mortal Draw and their own level is higher than that of a selected ISkillProvider, their own skill will be used anyway.