Resolves issue #20 - allow skills to process key input while not locked on
Added isLockedOn parameter to SkillActive::isKeyListener.

Skills that require locking on should always return false from this method when not locked on.

Skills that require the isLockedOn value for logic in the keyPressed or other methods are free to store the value from isKeyListener or retrieve it from the player as needed.